Usage of WaniKani vocabulary, new vocabulary page
Now WaniKani vocabulary is used to create reviews. Added prototype of a new vocabulary selection page. Did some refactoring and moved stuff into more appropriate places.
