+++ src/core/reviews/ReviewCreator.ts
@@ -1,0 +1,9 @@
+import type { Review } from "@/core/Review";
+
+interface ReviewCreator {
+  id: string;
+  name: string;
+  create(params: string[]): Review | undefined;
+}
+
+export default ReviewCreator;
